Move MainProductCard and AdditionalProductCard to components folder
Since they are used in more than one page it makes sense to indicate by their location that they're shared.
