+++ src/client/components/AdditionalProductCard.tsx
