+++ src/client/components/MainProductCard.tsx
